# docs/api.yml
openapi: 3.0.0
info:
  title: 'WeDigital Garden - Core API'
  description: 'Documentation of Core API for We Digital Garden'
  version: 0.0.1
servers:
  -
    url: 'http://localhost'
paths:
...
  '/api/expertises/{lang}':
    get:
      tags:
        - Expertise
      summary: 'Display a collection of expertises with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\ExpertiseController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A expertises collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/expertise_with_lang_collection_resource'
        '404':
          description: 'A expertises collection not found'
...
  '/api/expertises/{slug}':
    get:
      tags:
        - Expertise
      summary: 'Display the specified expertise resource.'
      operationId: 'App\Http\Controllers\Api\ExpertiseController::getOneBySlug'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: slug
          in: path
          description: 'String ID of the expertise to get'
          required: true
      responses:
        '200':
          description: 'One expertise'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/expertise_resource'
        '404':
          description: 'A expertise not found'
  '/api/expertises/{id}/{lang}':
    get:
      tags:
        - Expertise
      summary: 'Display the specified expertise with specific language resource.'
      operationId: 'App\Http\Controllers\Api\ExpertiseController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the office to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One expertise'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/expertise_with_lang_resource'
        '404':
          description: 'A expertise not found'
  '/api/expertises/{slug}/{lang}':
    get:
      tags:
        - Expertise
      summary: 'Display the specified expertise with specific language resource.'
      operationId: 'App\Http\Controllers\Api\ExpertiseController::getOneBySlugWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: slug
          in: path
          description: 'String ID of the office to get'
          required: true
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One expertise'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/expertise_with_lang_resource'
        '404':
          description: 'A expertise not found'
...
  '/api/formules/{lang}':
    get:
      tags:
        - Formule
      summary: 'Display a collection of formules with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\FormuleController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A formules collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/formule_with_lang_collection_resource'
        '404':
          description: 'A formules collection not found'
...
  '/api/formules/{id}/{lang}':
    get:
      tags:
        - Formule
      summary: 'Display the specified formule with specific language resource.'
      operationId: 'App\Http\Controllers\Api\FormuleController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the office to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One formule'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/formule_with_lang_resource'
        '404':
          description: 'A formule not found'
...
  '/api/framework/certifications/{lang}':
    get:
      tags:
        - Framework
      summary: 'Display a collection of Framework certifications with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\Framework\CertificationController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A Framework certifications collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/certification_with_lang_collection_resource'
        '404':
          description: 'A Framework certifications collection not found'
...
  '/api/framework/certifications/{id}/{lang}':
    get:
      tags:
        - Framework
      summary: 'Display the specified framework certification with specific language resource.'
      operationId: 'App\Http\Controllers\Api\Framework\CertificationController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the framework certification to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One framework certification'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/certification_with_lang_resource'
        '404':
          description: 'A framework certification not found'
...
  '/api/framework/values/{lang}':
    get:
      tags:
        - Framework
      summary: 'Display a collection of Framework values with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\Framework\ValueController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A Framework values collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/value_with_lang_collection_resource'
        '404':
          description: 'A Framework values collection not found'
...
  '/api/Framework values/{id}/{lang}':
    get:
      tags:
        - Framework
      summary: 'Display the specified framework value with specific language resource.'
      operationId: 'App\Http\Controllers\Api\Framework\ValueController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the framework value to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One framework value'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/value_with_lang_resource'
        '404':
          description: 'A framework value not found'
  /api/menus:
    get:
      tags:
        - Menu
      summary: 'Display a collection of menus ressources.'
      operationId: 'App\Http\Controllers\Api\MenuController::index'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
      responses:
        '200':
          description: 'A menus collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/menu_collection_resource'
        '404':
          description: 'A menus collection not found'
  '/api/menus/{lang}':
    get:
      tags:
        - Menu
      summary: 'Display a collection of menus with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\MenuController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A menu collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/menu_with_lang_collection_resource'
        '404':
          description: 'A menu collection not found'
  '/api/menus/{id}':
    get:
      tags:
        - Menu
      summary: 'Display the specified menu resource.'
      operationId: 'App\Http\Controllers\Api\MenuController::getOne'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the menu to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
      responses:
        '200':
          description: 'One menu'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/menu_resource'
        '404':
          description: 'A menu not found'
  '/api/menus/{id}/{lang}':
    get:
      tags:
        - Menu
      summary: 'Display the specified menu with specific language resource.'
      operationId: 'App\Http\Controllers\Api\MenuController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the menu to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One menu'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/menu_with_lang_resource'
        '404':
          description: 'A menu not found'
  '/api/menus/{slug}':
    get:
      tags:
        - Menu
      summary: 'Display the specified menu resource by slug.'
      operationId: 'App\Http\Controllers\Api\MenuController::getOneBySlug'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: slug
          in: path
          description: 'String ID of the menu to get'
          required: true
      responses:
        '200':
          description: 'One menu'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/menu_resource'
        '404':
          description: 'A menu not found'
  '/api/menus/{slug}/{lang}':
    get:
      tags:
        - Menu
      summary: 'Display the specified menu with specific language resource.'
      operationId: 'App\Http\Controllers\Api\MenuController::getOneBySlugWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: slug
          in: path
          description: 'String ID of the menu to get'
          required: true
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One menu'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/menu_with_lang_resource'
        '404':
          description: 'A menu not found'
...
  '/api/pages/{lang}':
    get:
      tags:
        - Page
      summary: 'Display a collection of pages with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\PageController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A pages collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page_with_lang_collection_resource'
        '404':
          description: 'A pages collection not found'
...
  '/api/pages/{slug}':
    get:
      tags:
        - Page
      summary: 'Display the specified page resource.'
      operationId: 'App\Http\Controllers\Api\PageController::getOneBySlug'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: slug
          in: path
          description: 'String ID of the page to get'
          required: true
      responses:
        '200':
          description: 'One page'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page_resource'
        '404':
          description: 'A page not found'
  '/api/pages/{id}/{lang}':
    get:
      tags:
        - Page
      summary: 'Display the specified page with specific language resource.'
      operationId: 'App\Http\Controllers\Api\PageController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the page to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One page'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page_with_lang_resource'
        '404':
          description: 'A page not found'
  '/api/pages/{slug}/{lang}':
    get:
      tags:
        - Page
      summary: 'Display the specified expertise with specific language resource.'
      operationId: 'App\Http\Controllers\Api\PageController::getOneBySlugWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: slug
          in: path
          description: 'String ID of the page to get'
          required: true
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One page'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/page_with_lang_resource'
        '404':
          description: 'A page not found'
...
  '/api/regions/{lang}':
    get:
      tags:
        - Region
      summary: 'Display a collection of regions with specific language ressources.'
      operationId: 'App\Http\Controllers\Api\RegionController::getAllWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'A regions collection'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/region_with_lang_collection_resource'
        '404':
          description: 'A regions collection not found'
...
  '/api/regions/{slug}':
    get:
      tags:
        - Region
      summary: 'Display the specified region resource.'
      operationId: 'App\Http\Controllers\Api\RegionController::getOneBySlug'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: slug
          in: path
          description: 'String ID of the region to get'
          required: true
      responses:
        '200':
          description: 'One region'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/region_resource'
        '404':
          description: 'A region not found'
  '/api/regions/{id}/{lang}':
    get:
      tags:
        - Region
      summary: 'Display the specified region with specific language resource.'
      operationId: 'App\Http\Controllers\Api\RegionController::getOneWithLang'
      parameters:
        -
          name: Content-Type
          in: header
          required: true
          schema:
            type: string
            example: application/json
        -
          name: X-Requested-With
          in: header
          required: true
          schema:
            type: string
            example: XMLHttpRequest
        -
          name: id
          in: path
          description: 'Numeric ID of the region to get'
          required: true
          schema:
            $ref: '#/components/schemas/resource_id'
        -
          name: lang
          in: path
          description: 'String for select a specific language'
          required: true
      responses:
        '200':
          description: 'One region'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/region_with_lang_resource'
        '404':
          description: 'A region not found'
...
components:
  schemas:
    resource_id:
      description: 'The unique identifier of a resource'
      minimum: 1
      example: '1'
...
    expertise_with_lang_collection_resource:
      required:
        - link
        - data
      properties:
        data:
          description: 'Transform the resource into an array.'
          type: array
          items:
            $ref: '#/components/schemas/expertise_with_lang_resource'
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/expertises/fr
          type: object
      type: object
...
    formule_with_lang_collection_resource:
      required:
        - link
        - data
      properties:
        data:
          description: 'Transform the resource into an array.'
          type: array
          items:
            $ref: '#/components/schemas/formule_with_lang_resource'
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/formules/fr
          type: object
      type: object
...
    value_with_lang_collection_resource:
      required:
        - link
        - data
      properties:
        data:
          description: 'Transform the resource into an array.'
          type: array
          items:
            $ref: '#/components/schemas/value_with_lang_resource'
        links:
          description: 'Transform the resource into an array.'
          required:
            - self
          properties:
            self:
              description: 'Transform the resource into an array.'
              type: string
              format: uri
              example: /api/framework/values/fr
          type: object
      type: object
...
  securitySchemes:
    bearer:
      type: http
      name: bearerAuth
      bearerFormat: JWT
      scheme: bearer